+++ sdk/storage/azure-storage-files-shares/src/share_directory_client.cpp
@@ -360,31 +360,4 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
         std::move(ret), std::move(result.RawResponse));
   }
 
-  ForceCloseAllDirectoryHandlesPagedResponse ShareDirectoryClient::ForceCloseAllHandles(
-      const ForceCloseAllDirectoryHandlesOptions& options,
-      const Azure::Core::Context& context) const
-  {
-    auto protocolLayerOptions = _detail::ShareRestClient::Directory::ForceCloseHandlesOptions();
-    protocolLayerOptions.HandleId = FileAllHandles;
-    if (options.ContinuationToken.HasValue() && !options.ContinuationToken.Value().empty())
-    {
-      protocolLayerOptions.ContinuationToken = options.ContinuationToken;
-    }
-    protocolLayerOptions.Recursive = options.Recursive;
-    auto response = _detail::ShareRestClient::Directory::ForceCloseHandles(
-        m_shareDirectoryUrl, *m_pipeline, context, protocolLayerOptions);
-
-    ForceCloseAllDirectoryHandlesPagedResponse pagedResponse;
-
-    pagedResponse.NumberOfHandlesClosed = response.Value.NumberOfHandlesClosed;
-    pagedResponse.NumberOfHandlesFailedToClose = response.Value.NumberOfHandlesFailedToClose;
-    pagedResponse.m_shareDirectoryClient = std::make_shared<ShareDirectoryClient>(*this);
-    pagedResponse.m_operationOptions = options;
-    pagedResponse.CurrentPageToken = options.ContinuationToken.ValueOr(std::string());
-    pagedResponse.NextPageToken = response.Value.ContinuationToken.ValueOr(std::string());
-    pagedResponse.RawResponse = std::move(response.RawResponse);
-
-    return pagedResponse;
-  }
-
 }}}} // namespace Azure::Storage::Files::Shares