+++ sdk/storage/azure-storage-files-shares/src/share_responses.cpp
@@ -79,22 +79,10 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     *this = m_shareFileClient->ListHandles(m_operationOptions, context);
   }
 
-  void ForceCloseAllFileHandlesPagedResponse::OnNextPage(const Azure::Core::Context& context)
-  {
-    m_operationOptions.ContinuationToken = std::move(NextPageToken);
-    *this = m_shareFileClient->ForceCloseAllHandles(m_operationOptions, context);
-  }
-
   void ListDirectoryHandlesPagedResponse::OnNextPage(const Azure::Core::Context& context)
   {
     m_operationOptions.ContinuationToken = std::move(NextPageToken);
     *this = m_shareDirectoryClient->ListHandles(m_operationOptions, context);
-  }
-
-  void ForceCloseAllDirectoryHandlesPagedResponse::OnNextPage(const Azure::Core::Context& context)
-  {
-    m_operationOptions.ContinuationToken = std::move(NextPageToken);
-    *this = m_shareDirectoryClient->ForceCloseAllHandles(m_operationOptions, context);
   }
 
 }}}} // namespace Azure::Storage::Files::Shares