+++ sdk/storage/azure-storage-files-shares/test/share_directory_client_test.cpp
@@ -441,11 +441,5 @@ namespace Azure { namespace Storage { namespace Test {
          pageResult.MoveToNextPage())
     {
     }
-    EXPECT_NO_THROW(m_fileShareDirectoryClient->ForceCloseAllHandles());
-    for (auto pageResult = m_fileShareDirectoryClient->ForceCloseAllHandles();
-         pageResult.HasMorePages();
-         pageResult.MoveToNextPage())
-    {
-    }
   }
 }}} // namespace Azure::Storage::Test